@@ -1,21 +1,17 @@
 apply plugin: 'com.android.application'
 apply plugin: 'realm-android'
-apply plugin: 'me.tatarka.retrolambda'
 
 apply from: 'script.gradle'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.earlier.yma"
-        minSdkVersion 15
-        targetSdkVersion 25
-        versionCode getVersionCode()
-        versionName "2.2"
-
-        multiDexEnabled true
-        testInstrumentationRunner "android.support.test.runner.AndroidUnitRunner"
+        minSdkVersion 21
+        targetSdkVersion 28
+        versionCode 45
+        versionName "2.21"
+        testInstrumentationRunner "androidx.test.runner.AndroidUnitRunner"
         resConfigs "en", "ko"
     }
     compileOptions {
