@@ -1,28 +1,26 @@
 package com.earlier.yma.utilities;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 /**
  * This provides methods to help Activities load their UI.
  */
 public class ActivityUtils {
 
-    /**
-     * The {@code fragment} is added to the container view with id {@code frameId}. The operation is
-     * performed by the {@code fragmentManager}.
-     */
-    public static void addFragmentToActivity(@NonNull FragmentManager fragmentManager,
-            @NonNull Fragment fragment, int frameId) {
-        checkNotNull(fragmentManager);
-        checkNotNull(fragment);
-        FragmentTransaction transaction = fragmentManager.beginTransaction();
-        transaction.add(frameId, fragment);
-        transaction.commit();
-    }
+  /**
+   * The {@code fragment} is added to the container view with id {@code frameId}. The operation is
+   * performed by the {@code fragmentManager}.
+   */
+  public static void addFragmentToActivity(@NonNull FragmentManager fragmentManager,
+      @NonNull Fragment fragment, int frameId) {
+    Utils.checkNotNull(fragmentManager);
+    Utils.checkNotNull(fragment);
+    FragmentTransaction transaction = fragmentManager.beginTransaction();
+    transaction.add(frameId, fragment);
+    transaction.commit();
+  }
 
 }
