@@ -1,20 +1,20 @@
 package com.earlier.yma;
 
 import android.content.Context;
-
 import dagger.Module;
 import dagger.Provides;
 
 @Module
 public class ApplicationModule {
-    private final Context mContext;
 
-    public ApplicationModule(Context context) {
-        this.mContext = context;
-    }
+  private final Context mContext;
 
-    @Provides
-    Context provideContext() {
-        return mContext;
-    }
+  public ApplicationModule(Context context) {
+    this.mContext = context;
+  }
+
+  @Provides
+  Context provideContext() {
+    return mContext;
+  }
 }
