@@ -2,6 +2,6 @@ package com.earlier.yma;
 
 public interface BasePresenter {
 
-    void start();
+  void start();
 
 }
