@@ -2,6 +2,6 @@ package com.earlier.yma;
 
 public interface BaseView<T> {
 
-    void setPresenter(T presenter);
+  void setPresenter(T presenter);
 
 }
