@@ -3,42 +3,41 @@ package com.earlier.yma.meal;
 import com.earlier.yma.BasePresenter;
 import com.earlier.yma.BaseView;
 import com.earlier.yma.data.Meal;
-
 import java.util.Date;
 
 public interface MealContract {
 
-    interface View extends BaseView<Presenter> {
+  interface View extends BaseView<Presenter> {
 
-        void showMeal(Meal meal);
+    void showMeal(Meal meal);
 
-        void showNoMeal(Date timestamp);
+    void showNoMeal(Date timestamp);
 
-        void showNetworkError();
+    void showNetworkError();
 
-        void showUnknownError();
+    void showUnknownError();
 
-        void showProgress();
+    void showProgress();
 
-        void showSetupDialog();
+    void showSetupDialog();
 
-        void updateTitle(Date date);
-    }
+    void updateTitle(Date date);
+  }
 
-    interface Presenter extends BasePresenter {
+  interface Presenter extends BasePresenter {
 
-        void loadData();
+    void loadData();
 
-        void setFiltering(MealFilterType filterType);
+    void setFiltering(MealFilterType filterType);
 
-        MealFilterType getFiltering();
+    MealFilterType getFiltering();
 
-        void setDate(Date date);
+    void setDate(Date date);
 
-        Date getDate();
+    Date getDate();
 
-        void destroy();
+    void destroy();
 
-    }
+  }
 
 }
