@@ -5,6 +5,6 @@ import dagger.Component;
 @Component(modules = SearchSchoolPresenterModule.class)
 public interface SearchSchoolComponent {
 
-    void inject(SearchSchoolActivity activity);
+  void inject(SearchSchoolActivity activity);
 
 }
