@@ -3,13 +3,14 @@ package com.earlier.yma.utilities;
 import io.realm.RealmObject;
 
 public class RealmString extends RealmObject {
-    private String value;
 
-    public String getValue() {
-        return value;
-    }
+  private String value;
 
-    public void setValue(String value) {
-        this.value = value;
-    }
+  public String getValue() {
+    return value;
+  }
+
+  public void setValue(String value) {
+    this.value = value;
+  }
 }
