@@ -3,7 +3,6 @@ package com.earlier.yma.utilities;
 import java.io.IOException;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
-
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
 import okhttp3.ResponseBody;
@@ -11,31 +10,34 @@ import retrofit2.Converter;
 import retrofit2.Retrofit;
 
 public class ToStringConverterFactory extends Converter.Factory {
-    private static final MediaType MEDIA_TYPE = MediaType.parse("text/plain");
 
-    @Override
-    public Converter<ResponseBody, ?> responseBodyConverter(Type type, Annotation[] annotations, Retrofit retrofit) {
-        if (String.class.equals(type)) {
-            return new Converter<ResponseBody, String>() {
-                @Override
-                public String convert(ResponseBody value) throws IOException {
-                    return value.string();
-                }
-            };
-        }
-        return null;
-    }
+  private static final MediaType MEDIA_TYPE = MediaType.parse("text/plain");
 
-    @Override
-    public Converter<?, RequestBody> requestBodyConverter(Type type, Annotation[] parameterAnnotations, Annotation[] methodAnnotations, Retrofit retrofit) {
-        if (String.class.equals(type)) {
-            return new Converter<String, RequestBody>() {
-                @Override
-                public RequestBody convert(String value) throws IOException {
-                    return RequestBody.create(MEDIA_TYPE, value);
-                }
-            };
+  @Override
+  public Converter<ResponseBody, ?> responseBodyConverter(Type type, Annotation[] annotations,
+      Retrofit retrofit) {
+    if (String.class.equals(type)) {
+      return new Converter<ResponseBody, String>() {
+        @Override
+        public String convert(ResponseBody value) throws IOException {
+          return value.string();
         }
-        return null;
+      };
     }
+    return null;
+  }
+
+  @Override
+  public Converter<?, RequestBody> requestBodyConverter(Type type,
+      Annotation[] parameterAnnotations, Annotation[] methodAnnotations, Retrofit retrofit) {
+    if (String.class.equals(type)) {
+      return new Converter<String, RequestBody>() {
+        @Override
+        public RequestBody convert(String value) throws IOException {
+          return RequestBody.create(MEDIA_TYPE, value);
+        }
+      };
+    }
+    return null;
+  }
 }
